requirements.txt
python-binance
TA-Lib
numpy
backtrader
flask
flask_cors